requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.1
flexpolyline==0.1.0
frozenlist==1.7.0
future==1.0.0
geocoder==1.38.1
geojson==3.2.0
h11==0.16.0
here-location-services==0.4.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.0
oauthlib==3.2.2
openai==1.93.0
openpyxl==3.1.5
openrouteservice==2.3.3
packaging==25.0
pandas==2.3.0
pillow==11.2.1
polyline==2.0.2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyhocon==0.3.61
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
ratelim==0.1.6
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tkintermapview==1.29
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.1
